feat: check command paths and filenames

+++ src/commands/file/delete.js
@@ -1,6 +1,7 @@
 import { access, unlink } from 'fs/promises';
 import { join, isAbsolute } from 'path';
 import * as directoryCommands from '../directory/index.js';
+import { fileCheck } from '../../utils/file-check.js';
 
 export const deleteFile = async (filePath) => {
     try {
@@ -12,6 +13,11 @@ export const deleteFile = async (filePath) => {
         const fullFilePath = isAbsolute(filePath) ? filePath : join(currentDirectory, filePath);
 
         await access(fullFilePath);
+        const isFile = await fileCheck(fullFilePath);
+        if (!isFile) {
+            throw new Error('Operation failed');
+        } 
+
         await unlink(fullFilePath);
         console.log('File was deleted!');
     } catch (err) {
